+++ app/controllers/paymentsCtrl.js
@@ -21,15 +21,17 @@ module.exports.displayAddNewPaymentOption = (req, res, next) => {
 // Posts new payment option for current user
 module.exports.addNewPaymentOption = (req, res, next) => {
   const { PaymentOption } = req.app.get('models');
+  let { type, account_number } = req.body;
   PaymentOption.create({
-    type: req.body.type,
-    account_number: req.body.account_number,
+    type,
+    account_number,
     customer_id: req.user.id,
     deleted: false
   })
-  .then( addedPayment => {
-    res.redirect('/payment/manage');
-  })
+    .then(addedPayment => {
+      res.redirect('/payment/manage');
+    })
+    .catch(err => next(err));
 };
 
 module.exports.removePaymentOption = (req, res, next) => {
@@ -41,10 +43,10 @@ module.exports.removePaymentOption = (req, res, next) => {
     }
   })
     .then(paymentToUpdate => {
-      return paymentToUpdate.updateAttributes({ deleted: true })
+      return paymentToUpdate.updateAttributes({ deleted: true });
     })
     .then(updatedPayment => {
-      res.json(updatedPayment);
+      res.status(200).json(updatedPayment);
     })
     .catch(err => {
       next(err);
